requirements.txt
streamlit
paho-mqtt
pandas